Simplified + refactored collector; Datasources are now responsible for materializing their facets
